Add YamlField.
Reorganize where custom fields are defined, and add a field to edit
YAML.
