+++ requirements.txt
@@ -17,5 +17,6 @@ PyMySQL==1.0.2
 pyScss==1.3.7
 python-dotenv==0.15.0
 pywebpack==1.2.0
+PyYAML==5.3.1
 requests==2.25.1
 SQLAlchemy==1.4.1
